+++ R/shapviz.R
@@ -94,7 +94,7 @@ shapviz.matrix = function(object, X, baseline = 0, collapse = NULL, ...) {
 #'
 #' # Will use numeric matrix "X_pred" as feature matrix
 #' x <- shapviz(fit, X_pred = X_pred)
-#' sv_importance(x, kind = "bar")
+#' sv_importance(x)
 #'
 #' # Will use original values as feature matrix
 #' x <- shapviz(fit, X_pred = X_pred, X = iris[, -1])